Crate platforms

Rust platform registry: provides programmatic access to information about valid Rust platforms

This crate provides an interface to the platform data canonically sourced from the Rust compiler:

https://doc.rust-lang.org/nightly/rustc/platform-support.html

Minimum Supported Rust Version

Rust 1.40 or higher.

Minimum supported Rust version can be changed in the future, but it will be done with a minor version bump.

Re-exports

pub use crate::platform::Platform;

Modules

platform

Rust platforms

target

Target cfg attributes. Documented in the “Conditional compilation” section of the Rust reference:

Structs

Error

Error type

PlatformReq

Platform requirements: glob-like expressions for matching Rust platforms as identified by a “target triple”, e.g. i686-apple-darwin.

Enums

Tier

Rust platform tiers: support levels are organized into three tiers, each with a different set of guarantees.

Constants

TARGET_ARCH

target_arch when building this crate: x86_64

TARGET_ENV

target_env when building this crate: gnu

TARGET_OS

target_os when building this crate: linux