Module platforms::platform

Rust platforms

Structs

Platform

Rust platforms supported by mainline rustc

PlatformReq

Platform requirements: glob-like expressions for matching Rust platforms as identified by a “target triple”, e.g. i686-apple-darwin.

Enums

Tier

Rust platform tiers: support levels are organized into three tiers, each with a different set of guarantees.

Constants

AARCH64_APPLE_DARWIN

aarch64-apple-darwin: Apple Silicon macOS (11+)

AARCH64_APPLE_IOS

aarch64-apple-ios: ARM64 iOS

AARCH64_APPLE_IOS_MACABI

aarch64-apple-ios-macabi: Apple Catalyst on ARM64

AARCH64_APPLE_IOS_SIM

aarch64-apple-ios-sim: Apple iOS Simulator on ARM64

AARCH64_APPLE_TVOS

aarch64-apple-tvos: ARM64 tvOS

AARCH64_BE_UNKNOWN_LINUX_GNU

aarch64_be-unknown-linux-gnu: ARM64 Linux (big-endian)

AARCH64_BE_UNKNOWN_LINUX_GNU_ILP32

aarch64_be-unknown-linux-gnu_ilp32: ARM64 Linux (big-endian, ILP32 ABI)

AARCH64_FUCHSIA

aarch64-fuchsia: ARM64 Fuchsia

AARCH64_LINUX_ANDROID

aarch64-linux-android: ARM64 Android

AARCH64_PC_WINDOWS_MSVC

aarch64-pc-windows-msvc: ARM64 MSVC (Windows 10)

AARCH64_UNKNOWN_FREEBSD

aarch64-unknown-freebsd: ARM64 FreeBSD

AARCH64_UNKNOWN_HERMIT

aarch64-unknown-hermit

AARCH64_UNKNOWN_LINUX_GNU

aarch64-unknown-linux-gnu: ARM64 Linux

AARCH64_UNKNOWN_LINUX_GNU_ILP32

aarch64-unknown-linux-gnu_ilp32: ARM64 Linux (ILP32 ABI)

AARCH64_UNKNOWN_LINUX_MUSL

aarch64-unknown-linux-musl: ARM64 Linux with MUSL

AARCH64_UNKNOWN_NETBSD

aarch64-unknown-netbsd

AARCH64_UNKNOWN_NONE

aarch64-unknown-none: Bare ARM64, hardfloat

AARCH64_UNKNOWN_NONE_SOFTFLOAT

aarch64-unknown-none-softfloat: Bare ARM64, softfloat

AARCH64_UNKNOWN_OPENBSD

aarch64-unknown-openbsd: ARM64 OpenBSD

AARCH64_UNKNOWN_REDOX

aarch64-unknown-redox: ARM64 Redox OS

AARCH64_UWP_WINDOWS_MSVC

aarch64-uwp-windows-msvc

AARCH64_WRS_VXWORKS

aarch64-wrs-vxworks

ARMEBV7R_NONE_EABI

armebv7r-none-eabi: Bare ARMv7-R, Big Endian

ARMEBV7R_NONE_EABIHF

armebv7r-none-eabihf: Bare ARMv7-R, Big Endian, hardfloat

ARMV4T_UNKNOWN_LINUX_GNUEABI

armv4t-unknown-linux-gnueabi

ARMV5TE_UNKNOWN_LINUX_GNUEABI

armv5te-unknown-linux-gnueabi: ARMv5TE Linux

ARMV5TE_UNKNOWN_LINUX_MUSLEABI

armv5te-unknown-linux-musleabi: ARMv5TE Linux with MUSL

ARMV5T_UNKNOWN_LINUX_UCLIBCEABI

armv5te-unknown-linux-uclibceabi: ARMv5TE Linux with uClibc

ARMV6_UNKNOWN_FREEBSD

armv6-unknown-freebsd: ARMv6 FreeBSD

ARMV6_UNKNOWN_NETBSD_EABIHF

armv6-unknown-netbsd-eabihf

ARMV7A_NONE_EABIHF

armv7a-none-eabihf: ARM Cortex-A, hardfloat

ARMV7S_APPLE_IOS

armv7s-apple-ios: ARMv7 iOS, Cortex-a9

ARMV7_APPLE_IOS

armv7-apple-ios: ARMv7 iOS, Cortex-a8

ARMV7_LINUX_ANDROIDEABI

armv7-linux-androideabi: ARMv7a Android

ARMV7_UNKNOWN_FREEBSD

armv7-unknown-freebsd: ARMv7 FreeBSD

ARMV7_UNKNOWN_LINUX_GNUEABI

armv7-unknown-linux-gnueabi: ARMv7 Linux

ARMV7_UNKNOWN_LINUX_GNUEABIHF

armv7-unknown-linux-gnueabihf: ARMv7 Linux, hardfloat

ARMV7_UNKNOWN_LINUX_MUSLEABI

armv7-unknown-linux-musleabi: ARMv7 Linux with MUSL

ARMV7_UNKNOWN_LINUX_MUSLEABIHF

armv7-unknown-linux-musleabihf: ARMv7 Linux with MUSL, hardfloat

ARMV7_UNKNOWN_NETBSD_EABIHF

armv7-unknown-netbsd-eabihf

ARMV7_WRS_VXWORKS_EABIHF

armv7-wrs-vxworks-eabihf

ARM_LINUX_ANDROIDEABI

arm-linux-androideabi: ARMv7 Android

ARM_UNKNOWN_LINUX_GNUEABI

arm-unknown-linux-gnueabi: ARMv6 Linux

ARM_UNKNOWN_LINUX_GNUEABIHF

arm-unknown-linux-gnueabihf: ARMv6 Linux, hardfloat

ARM_UNKNOWN_LINUX_MUSLEABI

arm-unknown-linux-musleabi: ARMv6 Linux with MUSL

ARM_UNKNOWN_LINUX_MUSLEABIHF

arm-unknown-linux-musleabihf: ARMv6 Linux, MUSL, hardfloat

ASMJS_UNKNOWN_EMSCRIPTEN

asmjs-unknown-emscripten: asm.js via Emscripten

I386_APPLE_IOS

i386-apple-ios: 32-bit x86 iOS

I586_PC_WINDOWS_MSVC

i586-pc-windows-msvc: 32-bit Windows w/o SSE

I586_UNKNOWN_LINUX_GNU

i586-unknown-linux-gnu: 32-bit Linux w/o SSE

I586_UNKNOWN_LINUX_MUSL

i586-unknown-linux-musl: 32-bit Linux w/o SSE, MUSL

I686_APPLE_DARWIN

i686-apple-darwin: 32-bit OSX (10.7+, Lion+)

I686_LINUX_ANDROID

i686-linux-android: 32-bit x86 Android

I686_PC_WINDOWS_GNU

i686-pc-windows-gnu: 32-bit MinGW (Windows 7+)

I686_PC_WINDOWS_MSVC

i686-pc-windows-msvc: 32-bit MSVC (Windows 7+)

I686_UNKNOWN_FREEBSD

i686-unknown-freebsd: 32-bit FreeBSD

I686_UNKNOWN_HAIKU

i686-unknown-haiku: 32-bit Haiku

I686_UNKNOWN_LINUX_GNU

i686-unknown-linux-gnu: 32-bit Linux (2.6.18+)

I686_UNKNOWN_LINUX_MUSL

i686-unknown-linux-musl: 32-bit Linux with MUSL

I686_UNKNOWN_NETBSD

i686-unknown-netbsd: NetBSD/i386 with SSE2

I686_UNKNOWN_OPENBSD

i686-unknown-openbsd: 32-bit OpenBSD

MIPS64EL_UNKNOWN_LINUX_GNUABI64

mips64el-unknown-linux-gnuabi64: MIPS64 (LE) Linux, n64 ABI

MIPS64EL_UNKNOWN_LINUX_MUSLABI64

mips64el-unknown-linux-muslabi64: MIPS64 (LE) Linux, n64 ABI, MUSL

MIPS64_UNKNOWN_LINUX_GNUABI64

mips64-unknown-linux-gnuabi64: MIPS64 Linux, n64 ABI

MIPS64_UNKNOWN_LINUX_MUSLABI64

mips64-unknown-linux-muslabi64: MIPS64 Linux, n64 ABI, MUSL

MIPSEL_UNKNOWN_LINUX_GNU

mipsel-unknown-linux-gnu: MIPS (LE) Linux

MIPSEL_UNKNOWN_LINUX_MUSL

mipsel-unknown-linux-musl: MIPS (LE) Linux with MUSL

MIPSEL_UNKNOWN_LINUX_UCLIBC

mipsel-unknown-linux-uclibc: MIPS (LE) Linux with uClibc

MIPS_UNKNOWN_LINUX_GNU

mips-unknown-linux-gnu: MIPS Linux

MIPS_UNKNOWN_LINUX_MUSL

mips-unknown-linux-musl: MIPS Linux with MUSL

MIPS_UNKNOWN_LINUX_UCLIBC

mips-unknown-linux-uclibc: MIPS Linux with uClibc

MSP430_NONE_ELF

msp430-none-elf: 16-bit MSP430 microcontrollers

NVPTX64_NVIDIA_CUDA

nvptx64-nvidia-cuda: --emit=asm generates PTX code that runs on NVIDIA GPUs

POWERPC64LE_UNKNOWN_LINUX_GNU

powerpc64le-unknown-linux-gnu: PPC64LE Linux

POWERPC64LE_UNKNOWN_LINUX_MUSL

powerpc64le-unknown-linux-musl

POWERPC64_UNKNOWN_LINUX_GNU

powerpc64-unknown-linux-gnu: PPC64 Linux

POWERPC64_UNKNOWN_LINUX_MUSL

powerpc64-unknown-linux-musl

POWERPC_UNKNOWN_LINUX_GNU

powerpc-unknown-linux-gnu: PowerPC Linux

POWERPC_UNKNOWN_LINUX_GNUSPE

powerpc-unknown-linux-gnuspe: PowerPC SPE Linux

POWERPC_UNKNOWN_LINUX_MUSL

powerpc-unknown-linux-musl

S390X_UNKNOWN_LINUX_GNU

s390x-unknown-linux-gnu: S390x Linux

S390X_UNKNOWN_LINUX_MUSL

s390x-unknown-linux-musl: S390x Linux (kernel 2.6.32, MUSL)

SPARC64_SUN_SOLARIS

sparcv9-sun-solaris: SPARC Solaris 10/11, illumos

SPARC64_UNKNOWN_LINUX_GNU

sparc64-unknown-linux-gnu: SPARC Linux

SPARC64_UNKNOWN_NETBSD

sparc64-unknown-netbsd: NetBSD/sparc64

SPARC_UNKNOWN_LINUX_GNU

sparc-unknown-linux-gnu: 32-bit SPARC Linux

THUMBV6M_NONE_EABI

thumbv6m-none-eabi: Bare Cortex-M0, M0+, M1

THUMBV7EM_NONE_EABI

thumbv7em-none-eabi: Bare Cortex-M4, M7

THUMBV7EM_NONE_EABIHF

thumbv7em-none-eabihf: Bare Cortex-M4F, M7F, FPU, hardfloat

THUMBV7M_NONE_EABI

thumbv7m-none-eabi: Bare Cortex-M3

THUMBV7NEON_LINUX_ANDROIDEABI

thumbv7neon-linux-androideabi: Thumb2-mode ARMv7a Android with NEON

THUMBV7NEON_UNKNOWN_LINUX_GNUEABIHF

thumbv7neon-unknown-linux-gnueabihf: Thumb2-mode ARMv7a Linux with NEON (kernel 4.4, glibc 2.23)

WASM_UNKNOWN_EMSCRIPTEN

wasm32-unknown-emscripten: WebAssembly via Emscripten

WASM_UNKNOWN_UNKNOWN

wasm32-unknown-unknown: WebAssembly

WASM_WASI

wasm32-wasi: WebAssembly with WASI

X86_64_APPLE_DARWIN

x86_64-apple-darwin: 64-bit OSX (10.7+, Lion+)

X86_64_APPLE_IOS

x86_64-apple-ios: 64-bit x86 iOS

X86_64_FORTANIX_UNKNOWN_SGX

x86_64-fortanix-unknown-sgx: Fortanix ABI for 64-bit Intel SGX

X86_64_FUCHSIA

x86_64-fuchsia: 64-bit x86 Fuchsia

X86_64_LINUX_ANDROID

x86_64-linux-android: 64-bit x86 Android

X86_64_PC_SOLARIS

x86_64-pc-solaris: 64-bit Solaris 10/11, illumos

X86_64_PC_WINDOWS_GNU

x86_64-pc-windows-gnu: 64-bit MinGW (Windows 7+)

X86_64_PC_WINDOWS_MSVC

x86_64-pc-windows-msvc: 64-bit MSVC (Windows 7+)

X86_64_SUN_SOLARIS

x86_64-sun-solaris: Deprecated target for 64-bit Solaris 10/11, illumos

X86_64_UNKNOWN_DRAGONFLY

x86_64-unknown-dragonfly: 64-bit DragonFlyBSD

X86_64_UNKNOWN_FREEBSD

x86_64-unknown-freebsd: 64-bit FreeBSD

X86_64_UNKNOWN_HAIKU

x86_64-unknown-haiku: 64-bit Haiku

X86_64_UNKNOWN_ILLUMOS

x86_64-unknown-illumos: illumos

X86_64_UNKNOWN_LINUX_GNU

x86_64-unknown-linux-gnu: 64-bit Linux (2.6.18+)

X86_64_UNKNOWN_LINUX_GNUX32

x86_64-unknown-linux-gnux32: 64-bit Linux

X86_64_UNKNOWN_LINUX_MUSL

x86_64-unknown-linux-musl: 64-bit Linux with MUSL

X86_64_UNKNOWN_NETBSD

x86_64-unknown-netbsd: NetBSD/amd64

X86_64_UNKNOWN_OPENBSD

x86_64-unknown-openbsd: 64-bit OpenBSD

X86_64_UNKNOWN_REDOX

x86_64-unknown-redox: Redox OS