Module platforms::platform::tier1

All Tier 1 Rust platforms All Tier 1 platforms supported by Rust. Sourced from:

https://forge.rust-lang.org/platform-support.html

Tier 1 platforms can be thought of as “guaranteed to work”. Specifically they will each satisfy the following requirements:

Official binary releases are provided for the platform.

Automated testing is set up to run tests for the platform.

Landing changes to the rust-lang/rust repository’s master branch is gated on tests passing.

Documentation for how to use and how to build the platform is available.

Constants

I686_APPLE_DARWIN
i686-apple-darwin: 32-bit OSX (10.7+, Lion+)

I686_PC_WINDOWS_GNU
i686-pc-windows-gnu: 32-bit MinGW (Windows 7+)

I686_PC_WINDOWS_MSVC
i686-pc-windows-msvc: 32-bit MSVC (Windows 7+)

I686_UNKNOWN_LINUX_GNU
i686-unknown-linux-gnu: 32-bit Linux (2.6.18+)

X86_64_APPLE_DARWIN
x86_64-apple-darwin: 64-bit OSX (10.7+, Lion+)

X86_64_PC_WINDOWS_GNU
x86_64-pc-windows-gnu: 64-bit MinGW (Windows 7+)

X86_64_PC_WINDOWS_MSVC
x86_64-pc-windows-msvc: 64-bit MSVC (Windows 7+)

X86_64_UNKNOWN_LINUX_GNU
x86_64-unknown-linux-gnu: 64-bit Linux (2.6.18+)