Module platforms::platform::tier2

All Tier 2 Rust platforms All Tier 2 platforms supported by Rust. Sourced from:

https://forge.rust-lang.org/platform-support.html

Tier 2 platforms can be thought of as “guaranteed to build”. Automated tests are not run so it’s not guaranteed to produce a working build, but platforms often work to quite a good degree and patches are always welcome!

Specifically, these platforms are required to have each of the following:

Official binary releases are provided for the platform.

Automated building is set up, but may not be running tests.

Landing changes to the rust-lang/rust repository’s master branch is gated on platforms building. For some platforms only the standard library is compiled, but for others rustc and cargo are too.

Constants

AARCH64_APPLE_IOS
aarch64-apple-ios: ARM64 iOS

AARCH64_FUCHSIA
aarch64-fuchsia: ARM64 Fuchsia

AARCH64_LINUX_ANDROID
aarch64-linux-android: ARM64 Android

AARCH64_UNKNOWN_LINUX_GNU
aarch64-unknown-linux-gnu: ARM64 Linux

AARCH64_UNKNOWN_LINUX_MUSL
aarch64-unknown-linux-musl: ARM64 Linux with MUSL

AARCH64_UNKNOWN_CLOUDABI
aarch64-unknown-cloudabi: ARM64 CloudABI

ARMV5TE_UNKNOWN_LINUX_GNUEABI
armv5te-unknown-linux-gnueabi: ARMv5TE Linux

ARMV7_APPLE_IOS
armv7-apple-ios: ARMv7 iOS, Cortex-a8

ARMV7_LINUX_ANDROIDEABI
armv7-linux-androideabi: ARMv7a Android

ARMV7_UNKNOWN_LINUX_GNUEABIHF
armv7-unknown-linux-gnueabihf: ARMv7 Linux

ARMV7_UNKNOWN_LINUX_MUSLEABIHF
armv7-unknown-linux-musleabihf: ARMv7 Linux with MUSL

ARMV7S_APPLE_IOS
armv7s-apple-ios: ARMv7 iOS, Cortex-a9

ARMV7_UNKNOWN_CLOUDABI_EABIHF
armv7-unknown-cloudabi-eabihf: ARMv7 CloudABI, hardfloat

ARM_LINUX_ANDROIDEABI
arm-linux-androideabi: ARMv7 Android

ARM_UNKNOWN_LINUX_GNUEABI
arm-unknown-linux-gnueabi: ARMv6 Linux

ARM_UNKNOWN_LINUX_GNUEABIHF
arm-unknown-linux-gnueabihf: ARMv6 Linux, hardfloat

ARM_UNKNOWN_LINUX_MUSLEABI
arm-unknown-linux-musleabi: ARMv6 Linux with MUSL

ARM_UNKNOWN_LINUX_MUSLEABIHF
arm-unknown-linux-musleabihf: ARMv6 Linux, MUSL, hardfloat

ASMJS_UNKNOWN_EMSCRIPTEN
asmjs-unknown-emscripten: asm.js via Emscripten

I386_APPLE_IOS
i386-apple-ios: 32-bit x86 iOS

I586_PC_WINDOWS_MSVC
i586-pc-windows-msvc: 32-bit Windows w/o SSE

I586_UNKNOWN_LINUX_GNU
i586-unknown-linux-gnu: 32-bit Linux w/o SSE

I586_UNKNOWN_LINUX_MUSL
i586-unknown-linux-musl: 32-bit Linux w/o SSE, MUSL

I686_LINUX_ANDROID
i686-linux-android: 32-bit x86 Android

I686_UNKNOWN_FREEBSD
i686-unknown-freebsd: 32-bit FreeBSD

I686_UNKNOWN_LINUX_MUSL
i686-unknown-linux-musl: 32-bit Linux with MUSL

I686_UNKNOWN_CLOUDABI
i686-unknown-cloudabi: 32-bit CloudABI

MIPS64EL_UNKNOWN_LINUX_GNUABI64
mips64el-unknown-linux-gnuabi64: MIPS64 (LE) Linux, n64 ABI

MIPS64_UNKNOWN_LINUX_GNUABI64
mips64-unknown-linux-gnuabi64: MIPS64 Linux, n64 ABI

MIPSEL_UNKNOWN_LINUX_GNU
mipsel-unknown-linux-gnu: MIPS (LE) Linux

MIPSEL_UNKNOWN_LINUX_MUSL
mipsel-unknown-linux-musl: MIPS (LE) Linux with MUSL

MIPS_UNKNOWN_LINUX_GNU
mips-unknown-linux-gnu: MIPS Linux

MIPS_UNKNOWN_LINUX_MUSL
mips-unknown-linux-musl: MIPS Linux with MUSL

POWERPC64_UNKNOWN_LINUX_GNU
powerpc64-unknown-linux-gnu: PPC64 Linux

POWERPC64LE_UNKNOWN_LINUX_GNU
powerpc64le-unknown-linux-gnu: PPC64LE Linux

POWERPC_UNKNOWN_LINUX_GNU
powerpc-unknown-linux-gnu: PowerPC Linux

POWERPC_UNKNOWN_LINUX_GNUSPE
powerpc-unknown-linux-gnuspe: PowerPC SPE Linux

S390X_UNKNOWN_LINUX_GNU
s390x-unknown-linux-gnu: S390x Linux

SPARC64_UNKNOWN_LINUX_GNU
sparc64-unknown-linux-gnu: SPARC Linux

SPARC64_SUN_SOLARIS
sparcv9-sun-solaris: SPARC Solaris 10/11, illumos

SPARC_UNKNOWN_LINUX_GNU
sparc-unknown-linux-gnu: 32-bit SPARC Linux

WASM_UNKNOWN_EMSCRIPTEN
wasm32-unknown-emscripten: WebAssembly via Emscripten

WASM_UNKNOWN_UNKNOWN
wasm32-unknown-unknown: WebAssembly

X86_64_APPLE_IOS
x86_64-apple-ios: 64-bit x86 iOS

X86_64_FUCHSIA
x86_64-fuchsia: 64-bit x86 Fuchsia

X86_64_LINUX_ANDROID
x86_64-linux-android: 64-bit x86 Android

X86_64_RUMPRUN_NETBSD
x86_64-rumprun-netbsd: 64-bit NetBSD Rump Kernel

X86_64_SUN_SOLARIS
x86_64-sun-solaris: 64-bit Solaris 10/11, illumos

X86_64_UNKNOWN_CLOUDABI
x86_64-unknown-cloudabi: 64-bit CloudABI

X86_64_UNKNOWN_FREEBSD
x86_64-unknown-freebsd: 64-bit FreeBSD

X86_64_UNKNOWN_LINUX_MUSL
x86_64-unknown-linux-musl: 64-bit Linux with MUSL

X86_64_UNKNOWN_NETBSD
x86_64-unknown-netbsd: NetBSD/amd64

X86_64_UNKNOWN_REDOX
x86_64-unknown-redox: Redox OS

X86_64_UNKNOWN_LINUX_GNUX32
x86_64-unknown-linux-gnux32: 64-bit Linux