Module platforms::target

Target cfg attributes. Documented in the “Conditional compilation” section of the Rust reference:

https://doc.rust-lang.org/reference/attributes.html#conditional-compilation

Enums

Arch

target_arch: Target CPU architecture

Env

target_env: target enviroment that disambiguates the target platform by ABI / libc. This value is closely related to the fourth element of the platform target triple, though it is not identical. For example, embedded ABIs such as gnueabihf will simply define target_env as "gnu" (i.e. target::Env::GNU)

OS

target_os: Operating system of the target. This value is closely related to the second and third element of the platform target triple, though it is not identical.

Constants

TARGET_ARCH

target_arch when building this crate: x86_64

TARGET_ENV

target_env when building this crate: gnu

TARGET_OS

target_os when building this crate: linux