Module platforms::platform

Rust platform types

Modules

tier1
All Tier 1 Rust platforms All Tier 1 platforms supported by Rust. Sourced from:

tier2
All Tier 2 Rust platforms All Tier 2 platforms supported by Rust. Sourced from:

tier3
All Tier 3 Rust platforms All Tier 3 platforms supported by Rust. Sourced from:

Structs

Platform
Rust platforms supported by mainline rustc

PlatformReq
Platform requirements: glob-like expressions for matching Rust platforms as identified by a "target triple", e.g. i686-apple-darwin.

Enums

Tier
Rust platform tiers: support levels are organized into three tiers, each with a different set of guarantees.

Constants

ALL_PLATFORMS
All valid Rust platforms usable from the mainline compiler