Module platforms::platform::tier3

All Tier 3 platforms supported by Rust. Sourced from:

https://doc.rust-lang.org/nightly/rustc/platform-support.html

Tier 3 platforms are those which the Rust codebase has support for, but which are not built or tested automatically, and may not work. Official builds are not available.

Constants

I686_UNKNOWN_HAIKU
i686-unknown-haiku: 32-bit Haiku

I686_UNKNOWN_NETBSD
i686-unknown-netbsd: NetBSD/i386 with SSE2

MIPSEL_UNKNOWN_LINUX_UCLIBC
mipsel-unknown-linux-uclibc: MIPS (LE) Linux with uClibc

MIPS_UNKNOWN_LINUX_UCLIBC
mips-unknown-linux-uclibc: MIPS Linux with uClibc

MSP430_NONE_ELF
msp430-none-elf: 16-bit MSP430 microcontrollers

SPARC64_UNKNOWN_NETBSD
sparc64-unknown-netbsd: NetBSD/sparc64

THUMBV6M_NONE_EABI
thumbv6m-none-eabi: Bare Cortex-M0, M0+, M1

THUMBV7EM_NONE_EABI
thumbv7em-none-eabi: Bare Cortex-M4, M7

THUMBV7EM_NONE_EABIHF
thumbv7em-none-eabihf: Bare Cortex-M4F, M7F, FPU, hardfloat

THUMBV7M_NONE_EABI
thumbv7m-none-eabi: Bare Cortex-M3

X86_64_FORTANIX_UNKNOWN_SGX
x86_64-fortanix-unknown-sgx: Fortanix ABI for 64-bit Intel SGX

X86_64_UNKNOWN_BITRIG
x86_64-unknown-bitrig: 64-bit Bitrig

X86_64_UNKNOWN_DRAGONFLY
x86_64-unknown-dragonfly: 64-bit DragonFlyBSD

X86_64_UNKNOWN_HAIKU
x86_64-unknown-haiku: 64-bit Haiku

X86_64_UNKNOWN_OPENBSD
x86_64-unknown-openbsd: 64-bit OpenBSD